Counts for gene

Function countsPerGene create a list with three data frames containing counts for gene or choosen gene part.

countsPerGene(datawt = list(), databrm = list(), datac = list(), gene,
  geneData = directRNAExplorer::TAIR10_genes, genePart = "gene",
  type = "unique")

Arguments

datawt
a list of bamDataFrames received after use the bamToR function containing information from the wt individual

databrm
a list of bamDataFrames received after use the bamToR function containing information from the brm individual

datac
a list of bamDataFrames received after use the bamToR function containing information from the c individual

gene
a character to indicate for which gene we compute the counts

geneData
a data.frame containing informations about genes

genePart
a character to indicate for what gene part we want to compute the counts, by default it is the "gene"

type
a character to indicate whether we count unique counts or all of them