genesSumarize

Function genesSumarize counts the coverage of each gene

genesSummarize(bamDataFrame, geneData = directRNAExplorer::TAIR10_genes_tidy,
  chromosome = NULL, range = 0, strand = "sense", genePart = "gene")

Arguments

bamDataFrame
Data frame converted to R using bamToR() function

geneData
Data frame with positions of all genes and their names, by default we use a TAIR10_genes_tidy

chromosome
Optional, number of chromosome

range
How many nucleotide before start and after stop we include to genes.

strand
On which strand we want to compute statistics, by default we use the sense strand, we can also compute statistics on the antisense strand.

genePart
Part of gene on which we compute the summary

Examples

library(directRNAExplorer)
data <- brmDataChromosome1[brmDataChromosome1$pos >2002000 & brmDataChromosome1$pos < 2018000,]
genesSummarize(data)

#>            name counts uniqueCounts length
#> 1     AT1G01010      0            0   2268
#> 2     AT1G01020      0            0   2809
#> 3     AT1G01030      0            0   2065
#> 4     AT1G01040      0            0   8081
#> 5     AT1G01046      0            0    206
#> 6     AT1G01050      0            0   1983
#> 7     AT1G01060      0            0   4492
#> 8     AT1G01070      0            0   2192
#> 9     AT1G01073      0            0    110
#> 10    AT1G01080      0            0   1723
#> 11    AT1G01090      0            0   1801
#> 12    AT1G01100      0            0   1124
#> 13    AT1G01110      0            0   2453
#> 14    AT1G01115      0            0    116
#> 15    AT1G01120      0            0   1898
#> 16    AT1G01130      0            0   1906
#> 17    AT1G01140      0            0   3459
#> 18    AT1G01150      0            0   2023
#> 19    AT1G01160      0            0   1757
#> 20    AT1G01170      0            0    806
#> 21    AT1G01180      0            0   1175
#> 22    AT1G01183      0            0    100
#> 23    AT1G01190      0            0   1819
#> 24    AT1G01200      0            0   1698
#> 25    AT1G01210      0            0    847
#> 26    AT1G01220      0            0   4275
#> 27    AT1G01225      0            0   1420
#> 28    AT1G01230      0            0   1784
#> 29    AT1G01240      0            0   1940
#> 30    AT1G01250      0            0    839
#> 31    AT1G01260      0            0   2663
#> 32    AT1G01270      0            0     71
#> 33    AT1G01280      0            0   1684
#> 34    AT1G01290      0            0   1822
#> 35    AT1G01300      0            0   1821
#> 36    AT1G01305      0            0    600
#> 37    AT1G01310      0            0    976
#> 38    AT1G01320      0            0   9510
#> 39    AT1G01340      0            0   3503
#> 40    AT1G01350      0            0   2038
#> 41    AT1G01355      0            0   1055
#> 42    AT1G01360      0            0   1212
#> 43    AT1G01370      0            0   2120
#> 44    AT1G01380      0            0    789
#> 45    AT1G01390      0            0   1686
#> 46    AT1G01400      0            0   1521
#> 47    AT1G01410      0            0   1085
#> 48    AT1G01420      0            0   1519
#> 49    AT1G01430      0            0   1854
#> 50    AT1G01440      0            0   2716
#> 51    AT1G01448      0            0   2820
#> 52    AT1G01450      0            0   1412
#> 53    AT1G01453      0            0   1253
#> 54    AT1G01460      0            0   2442
#> 55    AT1G01470      0            0    802
#> 56    AT1G01471      0            0    215
#> 57    AT1G01480      0            0   2618
#> 58    AT1G01490      0            0   2299
#> 59    AT1G01500      0            0   1790
#> 60    AT1G01510      0            0   2845
#> 61    AT1G01520      0            0   1543
#> 62    AT1G01530      0            0   1030
#> 63    AT1G01540      0            0   2904
#> 64    AT1G01550      0            0   2136
#> 65    AT1G01560      0            0   2199
#> 66    AT1G01570      0            0   2259
#> 67    AT1G01580      0            0   3646
#> 68    AT1G01590      0            0   3075
#> 69    AT1G01600      0            0   2292
#> 70    AT1G01610      0            0   2649
#> 71    AT1G01620      0            0   1637
#> 72    AT1G01630      0            0   1904
#> 73    AT1G01640      0            0   1514
#> 74    AT1G01650      0            0   4976
#> 75    AT1G01660      0            0   2551
#> 76    AT1G01670      0            0   3145
#> 77    AT1G01680      0            0   1956
#> 78    AT1G01690      0            0   3321
#> 79    AT1G01695      0            0   1548
#> 80    AT1G01700      0            0   1979
#> 81    AT1G01710      0            0   4943
#> 82    AT1G01720      0            0   1489
#> 83    AT1G01725      0            0    983
#> 84    AT1G01730      0            0   1105
#> 85    AT1G01740      0            0   2458
#> 86    AT1G01750      0            0    944
#> 87    AT1G01760      0            0   2182
#> 88    AT1G01770      0            0   4276
#> 89    AT1G01780      0            0   1484
#> 90    AT1G01790      0            0   6484
#> 91    AT1G01800      0            0   1698
#> 92    AT1G01810      0            0    638
#> 93    AT1G01820      0            0   2119
#> 94    AT1G01830      0            0   3780
#> 95    AT1G01840      0            0    821
#> 96    AT1G01860      0            0   2154
#> 97    AT1G01870      0            0     72
#> 98    AT1G01880      0            0   2629
#> 99    AT1G01890      0            0     72
#> 100   AT1G01900      0            0   2814
#> 101   AT1G01910      0            0   2801
#> 102   AT1G01920      0            0   3522
#> 103   AT1G01930      0            0   3024
#> 104   AT1G01940      0            0   1890
#> 105   AT1G01950      0            0   5240
#> 106   AT1G01960      0            0   6994
#> 107   AT1G01970      0            0   1786
#> 108   AT1G01980      0            0   1827
#> 109   AT1G01990      0            0   1175
#> 110   AT1G02000      0            0   1780
#> 111   AT1G02010      0            0   4589
#> 112   AT1G02020      0            0   2332
#> 113   AT1G02030      0            0   1937
#> 114   AT1G02040      0            0    974
#> 115   AT1G02050      0            0   1401
#> 116   AT1G02060      0            0   2176
#> 117   AT1G02065      0            0   2131
#> 118   AT1G02070      0            0    710
#> 119   AT1G02074      0            0    146
#> 120   AT1G02080      0            0  13512
#> 121   AT1G02090      0            0   2370
#> 122   AT1G02100      0            0   3004
#> 123   AT1G02110      0            0   2697
#> 124   AT1G02120      0            0   4136
#> 125   AT1G02130      0            0   1847
#> 126   AT1G02140      0            0   1266
#> 127   AT1G02145      0            0   4019
#> 128   AT1G02150      0            0   1989
#> 129   AT1G02160      0            0    849
#> 130   AT1G02170      0            0   1775
#> 131   AT1G02180      0            0   1123
#> 132   AT1G02190      0            0   2931
#> 133   AT1G02205      0            0   3690
#> 134   AT1G02210      0            0    263
#> 135   AT1G02220      0            0   1833
#> 136   AT1G02230      0            0   1602
#> 137   AT1G02250      0            0   1608
#> 138   AT1G02260      0            0   2436
#> 139   AT1G02270      0            0   3086
#> 140   AT1G02280      0            0   1963
#> 141   AT1G02290      0            0   2210
#> 142   AT1G02300      0            0   2175
#> 143   AT1G02305      0            0   2476
#> 144   AT1G02310      0            0   2563
#> 145   AT1G02320      0            0    146
#> 146   AT1G02330      0            0   1537
#> 147   AT1G02335      0            0   1069
#> 148   AT1G02340      0            0   2126
#> 149   AT1G02350      0            0    875
#> 150   AT1G02360      0            0   1170
#> 151   AT1G02370      0            0   2011
#> 152   AT1G02380      0            0   1971
#> 153   AT1G02390      0            0   2440
#> 154   AT1G02391      0            0    155
#> 155   AT1G02400      0            0   2776
#> 156   AT1G02405      0            0    526
#> 157   AT1G02410      0            0   2001
#> 158   AT1G02420      0            0   1475
#> 159   AT1G02430      0            0   1012
#> 160   AT1G02440      0            0    893
#> 161   AT1G02450      0            0    540
#> 162   AT1G02460      0            0   2531
#> 163   AT1G02470      0            0   1859
#> 164   AT1G02475      0            0   1561
#> 165   AT1G02480      0            0     72
#> 166   AT1G02490      0            0    116
#> 167   AT1G02500      0            0   2186
#> 168   AT1G02510      0            0   1170
#> 169   AT1G02520      0            0   5094
#> 170   AT1G02530      0            0   4882
#> 171   AT1G02540      0            0    872
#> 172   AT1G02550      0            0    728
#> 173   AT1G02560      0            0   2249
#> 174   AT1G02570      0            0    905
#> 175   AT1G02575      0            0    897
#> 176   AT1G02580      0            0   4419
#> 177   AT1G02590      0            0    623
#> 178   AT1G02600      0            0     71
#> 179   AT1G02610      0            0   2944
#> 180   AT1G02620      0            0    894
#> 181   AT1G02630      0            0   1814
#> 182   AT1G02640      0            0   3564
#> 183   AT1G02650      0            0   1714
#> 184   AT1G02660      0            0   3026
#> 185   AT1G02670      0            0   4253
#> 186   AT1G02680      0            0   1685
#> 187   AT1G02690      0            0   2959
#> 188   AT1G02700      0            0   1151
#> 189   AT1G02710      0            0    290
#> 190   AT1G02720      0            0   2410
#> 191   AT1G02730      0            0   4084
#> 192   AT1G02740      0            0   2622
#> 193   AT1G02750      0            0   1795
#> 194   AT1G02760      0            0     72
#> 195   AT1G02770      0            0   1820
#> 196   AT1G02780      0            0   1657
#> 197   AT1G02790      0            0   1933
#> 198   AT1G02800      0            0   2975
#> 199   AT1G02810      0            0   2215
#> 200   AT1G02813      0            0    611
#> 201   AT1G02816      0            0   1008
#> 202   AT1G02820      0            0    710
#> 203   AT1G02830      0            0    463
#> 204   AT1G02840      0            0   3078
#> 205   AT1G02850      0            0   2754
#> 206   AT1G02860      0            0   2118
#> 207   AT1G02870      0            0   1463
#> 208   AT1G02880      0            0   2151
#> 209   AT1G02890      0            0   6815
#> 210   AT1G02900      0            0    578
#> 211   AT1G02910      0            0   2571
#> 212   AT1G02920      0            0   1114
#> 213   AT1G02930      0            0   1279
#> 214   AT1G02940      0            0   1049
#> 215   AT1G02950      0            0   1300
#> 216   AT1G02952      0            0   1013
#> 217   AT1G02960      0            0   2965
#> 218   AT1G02965      0            0   1153
#> 219   AT1G02970      0            0   3072
#> 220   AT1G02980      0            0   3441
#> 221   AT1G02990      0            0   5919
#> 222   AT1G03000      0            0   4670
#> 223   AT1G03010      0            0   2902
#> 224   AT1G03020      0            0    308
#> 225   AT1G03022      0            0    125
#> 226   AT1G03030      0            0   2386
#> 227   AT1G03040      0            0   2725
#> 228   AT1G03050      0            0   2315
#> 229   AT1G03055      0            0   1785
#> 230   AT1G03060      0            0  14418
#> 231   AT1G03070      0            0   1656
#> 232   AT1G03080      0            0   6445
#> 233   AT1G03090      0            0   4505
#> 234   AT1G03100      0            0   2594
#> 235   AT1G03103      0            0    860
#> 236   AT1G03106      0            0    932
#> 237   AT1G03110      0            0   2717
#> 238   AT1G03120      0            0    869
#> 239   AT1G03130      0            0    891
#> 240   AT1G03140      0            0   1932
#> 241   AT1G03150      0            0   2198
#> 242   AT1G03160      0            0   4876
#> 243   AT1G03170      0            0   1052
#> 244   AT1G03180      0            0   1933
#> 245   AT1G03190      0            0   4535
#> 246   AT1G03200      0            0    683
#> 247   AT1G03210      0            0   1292
#> 248   AT1G03220      0            0   1529
#> 249   AT1G03230      0            0   1501
#> 250   AT1G03240      0            0    764
#>  [ reached getOption("max.print") -- omitted 28525 rows ]