matchToGene

Counts how many sequencing points were at known genes

matchToGene(positions, start, stop, geneName, strandData, strand = "sense",
  range = 0)

Arguments

positions
Positions from sequencing and informations about strand.

start
Gene start index.

stop
Gene stop index.

geneName
Gene name.

strandData
Vector with information that gene is on sense or antisense strand.

strand
On which strand we want to compute statistics, we can compute statistics on the sense or antisense strand.

range
How many nucleotide before start and after stop we include to genes.

Examples

# NOT RUN {
library(directRNAExplorer)
data <- dataChromosome1
dic <- TAIR10_genes
matchToGene(positions=unique(data$pos), start=dic$V4, stop=dic$V5, geneName=dic$id)
# }